+++ Package.swift
@@ -1,10 +1,13 @@
 import PackageDescription
 
 let package = Package(
-    name: "PromptLine",
-    dependencies: [
-        .Package(url: "https://github.com/kylef/PathKit.git", majorVersion: 0, minor: 7),
-        .Package(url: "https://github.com/antitypical/Result.git", majorVersion: 3),
-        .Package(url: "https://github.com/vknabel/Lens.git", majorVersion: 0, minor: 1),
-    ]
+  name: "PromptLine",
+  dependencies: [
+    .Package(url: "https://github.com/kylef/PathKit.git", majorVersion: 0, minor: 7),
+    .Package(url: "https://github.com/antitypical/Result.git", majorVersion: 3),
+    .Package(url: "https://github.com/vknabel/Lens.git", majorVersion: 0, minor: 1),
+
+    // Test Dependencies
+    .Package(url: "https://github.com/vknabel/Taps.git", majorVersion: 0, minor: 2)
+  ]
 )